[PaymentRequest] Move offline command and command handlers to PaymentBundle
