+++ src/Sylius/Bundle/PaymentBundle/spec/CommandHandler/Offline/CapturePaymentRequestHandlerSpec.php
@@ -11,20 +11,17 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Offline;
+namespace spec\Sylius\Bundle\PaymentBundle\CommandHandler\Offline;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\CoreBundle\PaymentRequest\Command\Offline\CapturePaymentRequest;
-use Sylius\Bundle\CoreBundle\PaymentRequest\Processor\Offline\CaptureProcessorInterface;
 use Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface;
-use Sylius\Component\Payment\Model\PaymentRequest;
+use Sylius\Bundle\PaymentBundle\Command\Offline\CapturePaymentRequest;
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
 
 final class CapturePaymentRequestHandlerSpec extends ObjectBehavior
 {
-    function let(
-        PaymentRequestProviderInterface $paymentRequestProvider
-    ): void {
+    function let(PaymentRequestProviderInterface $paymentRequestProvider): void
+    {
         $this->beConstructedWith($paymentRequestProvider);
     }
